Prevent ExtendedDaemonSet from being stuck because of an unhealthy node
In big clusters, it can happen that some nodes are not able to run a pod
either because they are starved in a given kind of resource or because the node is unreachable.
In those situations, the ExtendedDaemonSet should not wait for the pod to eventually start on the node.
It should instead proceed as if the failed node was not there.