Enable the possiblity to pause a Canary deployment
In the ExtendedDaemonset.Spec.Strategy, it is possible to define a
duration for the Canary deployment. At the end of this duration, if the
ExtendedDaemonset was not rolled back, the Canary deployment becomes
active and the rolling-update continues on all the nodes. By adding a
Paused field, it's now possible to prevent that from happening until
Paused is set back to false again. That's useful to investigate faulty
deployments as they currently are, without the need for a full rollback.

Another way we could've implemented this would be with an annotation,
but the last thing one needs when trying to prevent disaster is
misspelling an annotation that's not validated, so I preferred to go
with a new field.