Add a "Canary Paused" column to output
Here we add the column both for `kubectl get eds` (through changes to
the CRD) and to `kubectl eds get` (through changes to the plugin).
There's a small difference in the output whenever there is no canary, as
the plugin will output "-" instead of the empty string we get through
kubectl get, but since this behavior is also present for the Canary RS,
I'm not too worried.