Upgrade k8s.io/*, openapi definitions
This is an attempt to solve a tricky dependency version puzzle in
datadog-agent. Since that repo depends on this one (and datadog-operator,
and watermarkpodautoscaler) for type definitions, they need to be using
compatible openapi definitions. To get them, we also need to update
k8s.io to a recent enough version (but not too recent! dependency
issues).